refactor: espresso tokens being recomputed on every get call
- move computed property outside function to avoid recreating object again and again
